pyproject.toml
[project]
name = "iriam-point-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "easyocr>=1.7.2",
    "opencv-python>=4.11.0.86",
    "numpy>=2.2.5",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
